+++ lib/commanded/aggregates/aggregate.ex
@@ -69,28 +69,31 @@ defmodule Commanded.Aggregates.Aggregate do
     lifespan_timeout: :infinity
   ]
 
-  def start_link(config, args) do
-    aggregate_module = Keyword.fetch!(args, :aggregate_module)
-    aggregate_uuid = Keyword.fetch!(args, :aggregate_uuid)
+  def start_link(config, opts) do
+    {start_opts, aggregate_opts} =
+      Keyword.split(opts, [:debug, :name, :timeout, :spawn_opt, :hibernate_after])
 
-    unless is_atom(aggregate_module) and is_binary(aggregate_uuid) do
-      raise "aggregate_module must be an atom and aggregate_uuid must be a string"
-    end
+    aggregate_module = Keyword.fetch!(aggregate_opts, :aggregate_module)
+    aggregate_uuid = Keyword.fetch!(aggregate_opts, :aggregate_uuid)
+
+    unless is_atom(aggregate_module),
+      do: raise(ArgumentError, message: "aggregate module must be an atom")
+
+    unless is_binary(aggregate_uuid),
+      do: raise(ArgumentError, message: "aggregate identity must be a string")
 
     application = Keyword.fetch!(config, :application)
     snapshotting = Keyword.get(config, :snapshotting, %{})
     snapshot_options = Map.get(snapshotting, aggregate_module, [])
 
-    aggregate = %Aggregate{
+    state = %Aggregate{
       application: application,
       aggregate_module: aggregate_module,
       aggregate_uuid: aggregate_uuid,
       snapshotting: Snapshotting.new(application, aggregate_uuid, snapshot_options)
     }
 
-    opts = [name: args[:name]]
-
-    GenServer.start_link(__MODULE__, aggregate, opts)
+    GenServer.start_link(__MODULE__, state, start_opts)
   end
 
   @doc false
@@ -212,8 +215,8 @@ defmodule Commanded.Aggregates.Aggregate do
       end
 
     case lifespan_timeout do
-      :stop ->
-        {:stop, :normal, state}
+      {:stop, reason} ->
+        {:stop, reason, state}
 
       lifespan_timeout ->
         {:noreply, state, lifespan_timeout}
@@ -251,7 +254,7 @@ defmodule Commanded.Aggregates.Aggregate do
       {:reply, reply, state}
     else
       case lifespan_timeout do
-        :stop -> {:stop, :normal, reply, state}
+        {:stop, reason} -> {:stop, reason, reply, state}
         lifespan_timeout -> {:reply, reply, state, lifespan_timeout}
       end
     end
@@ -276,14 +279,14 @@ defmodule Commanded.Aggregates.Aggregate do
   @doc false
   @impl GenServer
   def handle_info({:events, events}, %Aggregate{} = state) do
-    %Aggregate{lifespan_timeout: lifespan_timeout} = state
+    %Aggregate{application: application, lifespan_timeout: lifespan_timeout} = state
 
     Logger.debug(fn -> describe(state) <> " received events: #{inspect(events)}" end)
 
     try do
       state =
         events
-        |> Upcast.upcast_event_stream()
+        |> Upcast.upcast_event_stream(additional_metadata: %{application: application})
         |> Enum.reduce(state, &handle_event/2)
 
       state = Enum.reduce(events, state, &handle_event/2)
@@ -391,44 +394,33 @@ defmodule Commanded.Aggregates.Aggregate do
     end
   end
 
-  # Rebuild aggregate state from a `Stream` of its events
+  # Rebuild aggregate state from a `Stream` of its events.
   defp rebuild_from_event_stream(event_stream, %Aggregate{} = state) do
-    %Aggregate{aggregate_module: aggregate_module} = state
+    Enum.reduce(event_stream, state, fn event, state ->
+      %RecordedEvent{data: data, stream_version: stream_version} = event
+      %Aggregate{aggregate_module: aggregate_module, aggregate_state: aggregate_state} = state
 
-    event_stream
-    |> Stream.map(fn event ->
-      {event.data, event.stream_version}
+      %Aggregate{
+        state
+        | aggregate_version: stream_version,
+          aggregate_state: aggregate_module.apply(aggregate_state, data)
+      }
     end)
-    |> Stream.transform(state, fn {event, stream_version}, state ->
-      case event do
-        nil ->
-          {:halt, state}
-
-        event ->
-          state = %Aggregate{
-            state
-            | aggregate_version: stream_version,
-              aggregate_state: aggregate_module.apply(state.aggregate_state, event)
-          }
-
-          {[state], state}
-      end
-    end)
-    |> Stream.take(-1)
-    |> Enum.at(0)
-    |> case do
-      nil -> state
-      state -> state
-    end
   end
 
-  defp aggregate_lifespan_timeout(lifespan, timeout_function_name, args) do
-    # Take the last event or the command/error
+  defp aggregate_lifespan_timeout(lifespan, function_name, args) do
+    # Take the last event or the command or error
     args = args |> List.wrap() |> Enum.take(-1)
 
-    case apply(lifespan, timeout_function_name, args) do
-      timeout when timeout in [:infinity, :hibernate, :stop] ->
+    case apply(lifespan, function_name, args) do
+      timeout when timeout in [:infinity, :hibernate] ->
         timeout
+
+      :stop ->
+        {:stop, :normal}
+
+      {:stop, _reason} = reply ->
+        reply
 
       timeout when is_integer(timeout) and timeout >= 0 ->
         timeout
@@ -437,7 +429,7 @@ defmodule Commanded.Aggregates.Aggregate do
         Logger.warn(fn ->
           "Invalid timeout for aggregate lifespan " <>
             inspect(lifespan) <>
-            ", expected a non-negative integer, `:infinity`, `:hibernate`, or `:stop` but got: " <>
+            ", expected a non-negative integer, `:infinity`, `:hibernate`, `:stop`, or `{:stop, reason}` but got: " <>
             inspect(invalid)
         end)
 
@@ -449,7 +441,7 @@ defmodule Commanded.Aggregates.Aggregate do
     %ExecutionContext{command: command, handler: handler, function: function} = context
     %Aggregate{aggregate_state: aggregate_state} = state
 
-    Logger.debug(fn -> describe(state) <> " executing command: #{inspect(command)}" end)
+    Logger.debug(fn -> describe(state) <> " executing command: " <> inspect(command) end)
 
     case Kernel.apply(handler, function, [aggregate_state, command]) do
       {:error, _error} = reply ->
@@ -473,6 +465,11 @@ defmodule Commanded.Aggregates.Aggregate do
       pending_events ->
         apply_and_persist_events(pending_events, context, state)
     end
+  rescue
+    error ->
+      Logger.error(fn -> Exception.format(:error, error, __STACKTRACE__) end)
+
+      {{:error, error}, state}
   end
 
   defp apply_and_persist_events(pending_events, context, %Aggregate{} = state) do