+++ test/commands/routing_commands_test.exs
@@ -1,5 +1,5 @@
 defmodule Commanded.Commands.RoutingCommandsTest do
-  use Commanded.StorageCase
+  use ExUnit.Case
 
   alias Commanded.DefaultApp
   alias Commanded.Commands.UnregisteredCommand
@@ -66,7 +66,7 @@ defmodule Commanded.Commands.RoutingCommandsTest do
 
   describe "routing to aggregate" do
     alias Commanded.Commands.AggregateRouter
-    alias Commanded.Commands.AggregateRoot.Command
+    alias Commanded.Commands.AggregateRoot.{Command, Command2}
 
     test "should dispatch command to registered handler" do
       assert :ok = AggregateRouter.dispatch(%Command{uuid: UUID.uuid4()}, @dispatch_opts)
@@ -75,6 +75,10 @@ defmodule Commanded.Commands.RoutingCommandsTest do
     test "should fail to dispatch unregistered command" do
       assert {:error, :unregistered_command} =
                AggregateRouter.dispatch(%UnregisteredCommand{}, @dispatch_opts)
+    end
+
+    test "should allow dispatching to a function other than execute/2" do
+      assert :ok = AggregateRouter.dispatch(%Command2{uuid: UUID.uuid4()}, @dispatch_opts)
     end
   end
 
@@ -212,26 +216,6 @@ defmodule Commanded.Commands.RoutingCommandsTest do
                  end
   end
 
-  test "should prevent registration for a command handler without a `handle/2` function" do
-    assert_raise ArgumentError,
-                 "Command handler `InvalidHandler` does not define a `handle/2` function",
-                 fn ->
-                   Code.eval_string("""
-                     alias Commanded.ExampleDomain.BankAccount
-                     alias Commanded.ExampleDomain.BankAccount.Commands.OpenAccount
-
-                     defmodule InvalidHandler do
-                     end
-
-                     defmodule InvalidRouter do
-                       use Commanded.Commands.Router
-
-                       dispatch OpenAccount, to: InvalidHandler, aggregate: BankAccount, identity: :account_number
-                     end
-                   """)
-                 end
-  end
-
   test "should show a helpful message when bad argument given to a `dispatch/2` function" do
     assert_raise RuntimeError,
                  """
@@ -247,57 +231,6 @@ defmodule Commanded.Commands.RoutingCommandsTest do
                        use Commanded.Commands.Router
 
                        dispatch OpenAccount, to: InvalidHandler, aggregate: BankAccount, id: :account_number
-                     end
-                   """)
-                 end
-  end
-
-  test "should prevent registrations for a invalid command module" do
-    assert_raise RuntimeError,
-                 "module `UnknownCommand` does not exist, perhaps you forgot to `alias` the namespace",
-                 fn ->
-                   Code.eval_string("""
-                     alias Commanded.ExampleDomain.BankAccount
-                     alias Commanded.ExampleDomain.OpenAccountHandler
-
-                     defmodule InvalidCommandRouter do
-                       use Commanded.Commands.Router
-
-                       dispatch UnknownCommand, to: OpenAccountHandler, aggregate: BankAccount, identity: :account_number
-                     end
-                   """)
-                 end
-  end
-
-  test "should prevent registrations for an invalid command handler module" do
-    assert_raise RuntimeError,
-                 "module `UnknownHandler` does not exist, perhaps you forgot to `alias` the namespace",
-                 fn ->
-                   Code.eval_string("""
-                     alias Commanded.ExampleDomain.BankAccount
-                     alias Commanded.ExampleDomain.BankAccount.Commands.OpenAccount
-
-                     defmodule InvalidHandlerRouter do
-                       use Commanded.Commands.Router
-
-                       dispatch OpenAccount, to: UnknownHandler, aggregate: BankAccount, identity: :account_number
-                     end
-                   """)
-                 end
-  end
-
-  test "should prevent registrations for an invalid aggregate module" do
-    assert_raise RuntimeError,
-                 "module `UnknownAggregate` does not exist, perhaps you forgot to `alias` the namespace",
-                 fn ->
-                   Code.eval_string("""
-                     alias Commanded.ExampleDomain.BankAccount.Commands.OpenAccount
-                     alias Commanded.ExampleDomain.OpenAccountHandler
-
-                     defmodule InvalidAggregateRouter do
-                       use Commanded.Commands.Router
-
-                       dispatch OpenAccount, to: OpenAccountHandler, aggregate: UnknownAggregate, identity: :account_number
                      end
                    """)
                  end