+++ test/commands/support/aggregate_root.ex
@@ -2,9 +2,12 @@ defmodule Commanded.Commands.AggregateRoot do
   @moduledoc false
   alias Commanded.Commands.AggregateRoot
 
-  defmodule Command, do: defstruct [uuid: nil]
+  defmodule Command, do: defstruct(uuid: nil)
+  defmodule Command2, do: defstruct(uuid: nil)
 
-  defstruct [uuid: nil]
+  defstruct uuid: nil
 
   def execute(%AggregateRoot{}, %Command{}), do: []
+
+  def custom_function_name(%AggregateRoot{}, %Command2{}), do: []
 end